+++ system/src/Grav/Common/Service/PagesServiceProvider.php
@@ -12,6 +12,7 @@ namespace Grav\Common\Service;
 use Grav\Common\Config\Config;
 use Grav\Common\Grav;
 use Grav\Common\Language\Language;
+use Grav\Common\Page\Markdown\MarkdownOutput;
 use Grav\Common\Page\Page;
 use Grav\Common\Page\Pages;
 use Grav\Common\Uri;
@@ -33,6 +34,7 @@ class PagesServiceProvider implements ServiceProviderInterface
     public function register(Container $container)
     {
         $container['pages'] = fn(Grav $grav) => new Pages($grav);
+        $container['markdown_output'] = fn(Grav $grav) => new MarkdownOutput($grav);
 
         if (defined('GRAV_CLI')) {
             $container['page'] = static function (Grav $grav) {